Материлаы 31-го урока

+++ Course/Lesson31/PracticeABC/EasyTrackerAPI/Model/Data/TrackerTask.cs
@@ -1,14 +1,15 @@
 using System;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json.Serialization;
 
 public class TrackerTask
 {  
-    public int ID { get; set; }
+    public long ID { get; set; }
     public string? Name { get; set; }
     public string? Description { get; set; }
     public bool IsComplete { get; set; } 
-    public DateTime? DueDate { get; set; }
-
-    public User? AssignedUser { get; set; }
-  
+    public DateTime? DueDate { get; set; } 
+     
+    public User? AssignedUser { get; set; } 
 }
